feat: Added notebook where parsing is done
Added a notebook where the parsing procedure is executed for all the
protocols provided in a certain folder. The protocol downloader will be
added in another notebook.

First all the protocols are parsed. In that case they are transformed
from html format to a csv file with two columns (speaker, speech)

The resulting csv files from the individual protocols are then combined
into one large csv file which has a third column describing the input
file. This way it is possible to know which speech was included in which
protocol