feat: Added protocol urls and a simple js file
The text file contains all the links to the currently available
html protocols of the Nationalrat. The javascript file contains a little
script that can be used to extract all the links pointing to the
protocols. It needs to be executed in the web console and the website
needs to be in the mode where all the protocols are shown on one page