feat: Added code (corpus creation from wikipedia)
The code in this commit is used to create a corpus of words, which are
relevant to a topic, by searching wikipedia for the topic, extracting
the text from there and then extract the words from the text by using
certain methods.

The inverted index is then used to search for the words in the speeches.